module_5/dual-model-genai-assistant/requirements.txt
python-dotenv==1.0.0
anthropic==0.7.1
openai==1.3.0
pydantic==2.5.0
